+++ traffic_portal/app/src/app.js
@@ -209,7 +209,6 @@ var trafficPortal = angular.module('trafficPortal', [
         require('./modules/private/types/deliveryServices').name,
         require('./modules/private/types/staticDnsEntries').name,
         require('./modules/private/users').name,
-        require('./modules/private/users/deliveryServices').name,
         require('./modules/private/users/edit').name,
         require('./modules/private/users/list').name,
         require('./modules/private/users/new').name,
@@ -394,7 +393,6 @@ var trafficPortal = angular.module('trafficPortal', [
         require('./common/modules/table/typeServers').name,
         require('./common/modules/table/typeStaticDnsEntries').name,
         require('./common/modules/table/users').name,
-        require('./common/modules/table/userDeliveryServices').name,
 
         // widgets
         require('./common/modules/widget/cacheGroups').name,